+++ docs/changelog.rst
@@ -18,6 +18,7 @@ Fixes
 * :class:`aspecd.processing.DatasetAlgebra` returns shape in error message if shapes differ.
 * Processing and analysis tasks issue warning if result
 * Ensure window length for Savitzky-Golay filter in :class:`aspecd.processing.Filtering` to always be odd
+* :class:`aspecd.processing.CommonRangeExtraction` ignores unit of last axis (*i.e.*, intensity) when checking for identical units
 
 
 Version 0.6.2